requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0
email-validator==2.0.0
loguru==0.7.2
